+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# v0.2.1
+
+`collections`
+* Refactoring of ExpiringStore
+
+`json`
+* Faster cloning in JsonTools
+* Add clone with TypeReference
+* Rename ambiguous calls
+
 # v0.1.3
 
 * Because of deployment tests. Code has not changed.